feat(launchpad): impl FocusManager and move table into separate component - simplifies alot of the old componenet draw/handle key event logic that   used custom `active` field on every component - uses FocusManager to instead deal with which component should be drawn   and should handle keystroke - split the status component into smaller components

+++ node-launchpad/src/mode.rs
@@ -24,7 +24,7 @@ pub enum Scene {
     StatusRewardsAddressPopUp,
     OptionsRewardsAddressPopUp,
     ManageNodesPopUp {
-        amount_of_nodes: usize,
+        amount_of_nodes: u64,
     },
     ResetNodesPopUp,
     UpgradeNodesPopUp,
